Reduce duplicate code by moving clustering to its own class
