+++ utils/collections.py
@@ -1,3 +1,4 @@
 import collections
 
 Location = collections.namedtuple('Location', ['lat', 'lng'])
+Relation = collections.namedtuple('Relation', ['other_event', 'distance', 'timedelta'])
